refactor: move getChildProcessEnv to break dependency loop
